Move video instruction modal to components folder
